Removed/consolidated TF override code
