+++ provider/terraform.rb
@@ -15,8 +15,6 @@ require 'provider/abstract_core'
 require 'provider/terraform/config'
 require 'provider/terraform/import'
 require 'provider/terraform/custom_code'
-require 'provider/terraform/property_override'
-require 'provider/terraform/resource_override'
 require 'provider/overrides/terraform/resource_override'
 require 'provider/overrides/terraform/property_override'
 require 'provider/terraform/sub_template'
